style: UI 자잘한 스타일 정리 및 모달 스크롤 잠금 적용
