+++ src/shared/styles/globals.css
@@ -1,1 +1,5 @@
 @import 'tailwindcss';
+
+body {
+  background-color: #191919;
+}
