+++ src/shared/ui/SuspenseFallback.tsx
