+++ src/widgets/header/ui/Header.tsx
@@ -117,7 +117,7 @@ const Header = ({ initialUserInfoData }: HeaderProps) => {
 
   return (
     <header
-      className={cn('sticky top-0 z-50 flex h-18 items-center justify-between bg-[#191919] px-10')}
+      className={cn('sticky top-0 flex h-18 items-center justify-between bg-[#191919] px-10')}
     >
       <Link href="/">
         <Logo />
